% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn_bp.R
\name{learn_bp11}
\alias{learn_bp11}
\title{Learn Back Propagation in 11 lines}
\usage{
learn_bp11(X, y)
}
\arguments{
\item{X}{input data}

\item{y}{output data}
}
\description{
Learn Back Propagation in 11 lines
}
\examples{
# construct new data
X = matrix(c(0,0,1,0,1,1,1,0,1,1,1,1), nrow=4, byrow=TRUE)
y = matrix(c(0,1,1,0),nrow=4)

# run 11 lines function
learn_bp11(X, y)

# view output
syn0
syn1
}
\references{
\url{http://qua.st/handcoding-neural-network/}
}

