% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn_do.R
\name{learn_do15}
\alias{learn_do15}
\title{Learn Back Propagation in 11 lines}
\usage{
learn_do15(X, y, alpha, hidden_dim, dropout_percent, do_dropout = TRUE)
}
\arguments{
\item{X}{input data}

\item{y}{output data}

\item{alpha}{alpha to be used}

\item{hidden_dim}{dimension of the hidden layer}

\item{dropout_percent}{percentage to be used for the dropout}

\item{do_dropout}{should dropout be used}
}
\description{
Learn Back Propagation in 11 lines
}
\examples{
# construct data
X = matrix(c(0,0,1,0,1,1,1,0,1,1,1,1), nrow=4, byrow=TRUE)
y = matrix(c(0,1,1,0),nrow=4)

# set hyperparameters
alpha = 0.5
hidden_dim = 4
dropout_percent = 0.2
do_dropout = TRUE

# run 11 lines function
learn_do15(X, y, alpha, hidden_dim, dropout_percent, TRUE)

# view output
synapse_0
synapse_1
}
\references{
\url{http://qua.st/handcoding-dropout/}
}

