% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/learn_gd.R
\name{learn_gd13}
\alias{learn_gd13}
\title{Learn Gradient Descent in 13 lines}
\usage{
learn_gd13(X, y, alpha, hidden_dim)
}
\arguments{
\item{X}{input data}

\item{y}{output data}

\item{alpha}{alpha to be used}

\item{hidden_dim}{dimension of the hidden layer}
}
\description{
Learn Gradient Descent in 13 lines
}
\examples{
# create new data
alpha = 0.5
hidden_dim = 4
X = matrix(c(0,0,1,0,1,1,1,0,1,1,1,1), nrow=4, byrow=TRUE)
y = matrix(c(0,1,1,0),nrow=4)

# run 13 lines function
learn_gd13(X, y, alpha, hidden_dim)
}
\references{
\url{http://qua.st/handcoding-gradient-descent/}
\url{http://iamtrask.github.io/2015/07/27/python-network-part2/}
}
\seealso{
\code{\link{learn_gd}}
}

