% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learNN.R
\docType{package}
\name{learnn}
\alias{learnn}
\title{Learn Neural Networks}
\description{
Implementations of several basic neural network concepts in R,
as based on posts on \url{https://qua.st/}.
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
